+++ plugins/core/qRANSAC_SD/RANSAC_SD_orig/Cylinder.h
@@ -91,7 +91,7 @@ private:
 			ScalarType *values, ScalarType *temp) const
 		{
 			ScalarType chi = 0;
-			int size = end - begin;
+			size_t size = end - begin;
 #ifdef DOPARALLEL
 			#pragma omp parallel for schedule(static) reduction(+:chi)
 #endif
@@ -117,7 +117,7 @@ private:
 		void Derivatives(const ScalarType *params, IteratorT begin, IteratorT end,
 			const ScalarType *values, const ScalarType *temp, ScalarType *matrix) const
 		{
-			int size = end - begin;
+			size_t size = end - begin;
 #ifdef DOPARALLEL
 			#pragma omp parallel for schedule(static)
 #endif
