+++ plugins/core/qRANSAC_SD/RANSAC_SD_orig/Torus.cpp
@@ -429,7 +429,7 @@ public:
 		ScalarType *values, ScalarType *temp) const
 	{
 		ScalarType chi = 0;
-		int size = end - begin;
+		size_t size = end - begin;
 #ifdef DOPARALLEL
 		#pragma omp parallel for schedule(static) reduction(+:chi)
 #endif
@@ -460,7 +460,7 @@ public:
 	void Derivatives(const ScalarType *params, IteratorT begin, IteratorT end,
 		const ScalarType *values, const ScalarType *temp, ScalarType *matrix) const
 	{
-		int size = end - begin;
+		size_t size = end - begin;
 #ifdef DOPARALLEL
 		#pragma omp parallel for schedule(static)
 #endif
